fix(simple-combo): adding loc to the base dir instead
