+++ projects/igniteui-angular/src/lib/combo/combo.common.ts
@@ -40,6 +40,8 @@ import {
     IgxComboFooterDirective, IgxComboHeaderDirective, IgxComboHeaderItemDirective, IgxComboItemDirective, IgxComboToggleIconDirective
 } from './combo.directives';
 import { IComboItemAdditionEvent, IComboSearchInputEventArgs } from './public_api';
+import { IComboResourceStrings } from '../core/i18n/combo-resources';
+import { CurrentResourceStrings } from '../core/i18n/resources';
 
 export const IGX_COMBO_COMPONENT = new InjectionToken<IgxComboBase>('IgxComboComponentToken');
 
@@ -441,6 +443,23 @@ export abstract class IgxComboBaseDirective extends DisplayDensityBase implement
 
     public set type(val: IgxInputGroupType) {
         this._type = val;
+    }
+
+    /**
+     * Gets/Sets the resource strings.
+     *
+     * @remarks
+     * By default it uses EN resources.
+     */
+    @Input()
+    public get resourceStrings(): IComboResourceStrings {
+        if (!this._resourceStrings) {
+            this._resourceStrings = CurrentResourceStrings.ComboResStrings;
+        }
+        return this._resourceStrings;
+    }
+    public set resourceStrings(value: IComboResourceStrings) {
+        this._resourceStrings = Object.assign({}, this._resourceStrings, value);
     }
 
     /**
@@ -890,6 +909,7 @@ export abstract class IgxComboBaseDirective extends DisplayDensityBase implement
     protected _filteredData = [];
     protected _displayKey: string;
     protected _remoteSelection = {};
+    protected _resourceStrings;
     protected _valid = IgxComboState.INITIAL;
     protected ngControl: NgControl = null;
     protected destroy$ = new Subject<any>();
