+++ projects/igniteui-angular/src/lib/simple-combo/simple-combo.component.html
@@ -98,7 +98,7 @@
     {{display}}
 </ng-template>
 <ng-template #empty>
-    <span>The list is empty</span>
+    <span>{{resourceStrings.igx_combo_empty_message}}</span>
 </ng-template>
 <ng-template #addItemDefault let-control>
     <button igxButton="flat" igxRipple>Add item</button>
